+++ src/side/right/selectedCharacterCancellableByPlayerTwo.presenter.jsx
@@ -2,6 +2,7 @@ import React, { useEffect } from "react";
 import useInput from "../../input/useInputPlayerTwo.hook";
 import { B } from "../../input/event";
 import useCharacterName from "../../character/useCharacterName.hook";
+import useCharacterAuthor from "../../character/useCharacterAuthor.hook";
 import useNavigationDispatch from "../../navigation/useDispatch.hook";
 import useCancelSound from "../../configuration/useCancelSound.hook";
 import unselectCharacterTwo from "../../navigation/action/unselectCharacterTwo.action";
@@ -9,12 +10,14 @@ import useColorIndex from "./useColorIndex.hook";
 import Portrait from "./portrait.view";
 import StandAnimation from "./standAnimation.view";
 import CharacterName from "./characterName.view";
+import CharacterAuthor from "./characterAuthor.view";
 import Type from "./type.view";
 
 export default function SelectedCharacterCancellableByPlayerTwo({ character }) {
   const dispatch = useNavigationDispatch();
   const input = useInput();
   const characterName = useCharacterName(character);
+  const characterAuthor = useCharacterAuthor(character);
   const cancelSound = useCancelSound();
   const colorIndex = useColorIndex();
 
@@ -36,6 +39,7 @@ export default function SelectedCharacterCancellableByPlayerTwo({ character }) {
       <Portrait character={character} />
       <StandAnimation character={character} colorIndex={colorIndex} />
       <CharacterName>{characterName}</CharacterName>
+      <CharacterAuthor>{characterAuthor}</CharacterAuthor>
       <Type>Player 2</Type>
     </>
   );