+++ src/side/right/selectingCharacterStyleByPlayerOne.presenter.jsx
@@ -2,6 +2,7 @@ import React, { useEffect } from "react";
 import useInput from "../../input/useInputPlayerOne.hook";
 import { A, B } from "../../input/event";
 import useCharacterName from "../../character/useCharacterName.hook";
+import useCharacterAuthor from "../../character/useCharacterAuthor.hook";
 import useCharacterStyleNames from "../../character/useCharacterStyleNames.hook";
 import useCharacterStyleIndex from "../../character/useCharacterStyleIndex.hook";
 import useCharacterStyle from "../../character/useCharacterStyle.hook";
@@ -16,6 +17,7 @@ import StyleSelector from "../../character/styleSelector.view";
 import Portrait from "./portrait.view";
 import StandAnimation from "./standAnimation.view";
 import CharacterName from "./characterName.view";
+import CharacterAuthor from "./characterAuthor.view";
 import Type from "./type.view";
 import Zone from "./zone.view";
 
@@ -38,6 +40,7 @@ export default function SelectingCharacterStyleByPlayerOne({ character }) {
   const characterStyleIndex = useCharacterStyleIndex(input, styleNames.length, onChange, defaultStyleIndex);
   const characterStyle = useCharacterStyle(character, characterStyleIndex);
   const characterName = useCharacterName(characterStyle);
+  const characterAuthor = useCharacterAuthor(characterStyle);
 
   useEffect(() => {
     const onCancel = () => {
@@ -71,6 +74,7 @@ export default function SelectingCharacterStyleByPlayerOne({ character }) {
       </Zone>
       <StandAnimation character={characterStyle} />
       <CharacterName>{characterName}</CharacterName>
+      <CharacterAuthor>{characterAuthor}</CharacterAuthor>
       <Type>Computer</Type>
     </>
   );